fix: Update etc_cost calculation in ERP macros
- Modified the calculation of 'etc_cost' to include 'expected_payout' along with 'service_fee' and 'delv_cost' across multiple ERP macro files.
- Changed the data type conversion of 'etc_cost' to first convert to integer before converting to string for consistency.

+++ utils/macros/ERP/v3/g_a_erp_macro_v3.py
@@ -31,7 +31,7 @@ class GAERPMacroV3:
         # 금액 계산
         df['etc_cost'] = df['expected_payout'].fillna(
             0) + df['service_fee'].fillna(0) + df['delv_cost'].fillna(0)
-        df['etc_cost'] = df['etc_cost'].astype(str)
+        df['etc_cost'] = df['etc_cost'].astype(int).astype(str)
         
         # 스타배송 평균 배송비 적용
         if self.is_star:
